% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ascii_world_map}
\alias{ascii_world_map}
\title{ascii_world_map}
\usage{
ascii_world_map(lon = c(), lat = c())
}
\arguments{
\item{lon}{test}

\item{lat}{test}
}
\description{
plot world map on the console
}
