% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_option_to_pandora_table}
\alias{convert_option_to_pandora_table}
\title{convert_option_to_pandora_table}
\usage{
convert_option_to_pandora_table(entity_type = NA)
}
\arguments{
\item{entity_type}{a valid entity type (e.g. site, sample, individual etc.)}
}
\description{
Given a command line `-t` options, retrieve the corresponding PANDORA table name and (full) ID column
}
