% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_module.R
\name{examples_module}
\alias{examples_module}
\title{examples_module}
\usage{
examples_module(quit_after = T)
}
\arguments{
\item{quit_after}{test}
}
\description{
This module gives you a quick overview of the sidora.cli interface
}
