% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fuzzy_search}
\alias{fuzzy_search}
\title{fuzzy_search}
\usage{
fuzzy_search(pattern, search_vector)
}
\arguments{
\item{pattern}{test}

\item{search_vector}{test}
}
\description{
fuzzy text search for entities to provide alternative suggestions
}
