% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_module.R
\name{glance_module}
\alias{glance_module}
\title{glance_module}
\usage{
glance_module(
  con = sidora.core::get_pandora_connection(),
  entity_type,
  cache_dir = "~/.sidora"
)
}
\arguments{
\item{con}{test}

\item{entity_type}{test}

\item{cache_dir}{test}
}
\description{
This module gives you an overview about the variables (columns) in a Pandora table.
}
