% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list}
\alias{list}
\title{list}
\usage{
list(con, entity_type, cache_dir)
}
\arguments{
\item{con}{a pandora connection from sidora.core}

\item{entity_type}{a pandora table to the generate a list of IDs from. Options: site,}

\item{cache_dir}{location of your cache directory.}
}
\description{
A module where, given a pandora entity type, a row-wise list of IDs will be printed for screen. Useful for fast look up of certain IDs.
}
\examples{
list(con, "site", cache_dir)
}
