% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_module.R
\name{list_module}
\alias{list_module}
\title{list_module}
\usage{
list_module(
  con = sidora.core::get_pandora_connection(),
  entity_type,
  entity_id = c(),
  filter_entity_type = NA,
  filter_string = NA,
  cache_dir = "~/.sidora"
)
}
\arguments{
\item{con}{a pandora connection from sidora.core}

\item{entity_type}{a pandora table to the generate a list of IDs from. Options: site,}

\item{cache_dir}{location of your cache directory.}
}
\description{
A module where, given a pandora entity type, a row-wise list of IDs will be printed for screen. Useful for fast look up of certain IDs.
}
