% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_joint_table}
\alias{make_joint_table}
\title{Get the joint table from Pandora, restricted to a specific project}
\usage{
make_joint_table(con, pr)
}
\arguments{
\item{con}{The DB connection object}

\item{pr}{The name of the Project}
}
\value{
A tibble joining sites, individuals, samples,
  extracts and libraries
}
\description{
Get the joint table from Pandora, restricted to a specific project
}
