% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_progress_table}
\alias{make_progress_table}
\title{Make an individual-based progress table}
\usage{
make_progress_table(joint_table)
}
\arguments{
\item{joint_table}{The joint table object generated with \code{\link{make_joint_table}}}
}
\value{
A tibble containing the progress table.
}
\description{
Make an individual-based progress table
}
