% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_extract.R, R/summarise_individual.R,
%   R/summarise_library.R, R/summarise_sample.R, R/summarise_site.R
\name{summarise_extract}
\alias{summarise_extract}
\alias{summarise_individual}
\alias{summarise_library}
\alias{summarise_sample}
\alias{summarise_site}
\title{summarise_site}
\usage{
summarise_extract(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001.A01",
  cache_dir = "~/.sidora"
)

summarise_individual(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001",
  cache_dir = "~/.sidora"
)

summarise_library(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001.A0101",
  cache_dir = "~/.sidora"
)

summarise_sample(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001.A",
  cache_dir = "~/.sidora"
)

summarise_site(
  con = sidora.core::get_pandora_connection(),
  entity_id = "Futuna",
  cache_dir = "~/.sidora"
)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_site
}
