% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_extract.R
\name{summarise_extract}
\alias{summarise_extract}
\title{summarise_extract}
\usage{
summarise_extract(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001.A01",
  cache_dir = "/tmp/sidora.cli_table_cache"
)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_extract
}
