% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_individual.R
\name{summarise_individual}
\alias{summarise_individual}
\title{summarise_individual}
\usage{
summarise_individual(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001",
  cache_dir = "/tmp/sidora.cli_table_cache"
)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_individual
}
