% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_library.R
\name{summarise_library}
\alias{summarise_library}
\title{summarise_library}
\usage{
summarise_library(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001.A0101",
  cache_dir = "/tmp/sidora.cli_table_cache"
)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_library
}
