% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_sample.R
\name{summarise_sample}
\alias{summarise_sample}
\title{summarise_sample}
\usage{
summarise_sample(
  con = sidora.core::get_pandora_connection(),
  entity_id = "FUT001.A",
  cache_dir = "/tmp/sidora.cli_table_cache"
)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_sample
}
