% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_site.R
\name{summary_site}
\alias{summary_site}
\title{summary_site}
\usage{
summary_site(con, entity_id, cache_dir)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summary_site
}
