% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_module.R
\name{tutorial_module}
\alias{tutorial_module}
\title{tutorial_module}
\usage{
tutorial_module(con, cache_dir)
}
\arguments{
\item{con}{test}

\item{cache_dir}{test}
}
\description{
This module guides you through a series of example applications
}
