% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_module.R
\name{view_module}
\alias{view_module}
\title{view_module}
\usage{
view_module(con, entity_type, entity_id, cache_dir)
}
\arguments{
\item{con}{test}

\item{entity_type}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
This module can be seen as a command-line replacement for the PANDORA web UI for viewing information about a single object.
}
