% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_individual.R
\name{summarise_individual}
\alias{summarise_individual}
\title{summarise_individual}
\usage{
summarise_individual(con, entity_id, cache_dir)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_individual
}
