% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_site.R
\name{summarise_site}
\alias{summarise_site}
\title{summarise_site}
\usage{
summarise_site(
  con = sidora.core::get_pandora_connection(),
  entity_id = "Futuna",
  cache_dir = "/tmp/sidora.cli_table_cache"
)
}
\arguments{
\item{con}{test}

\item{entity_id}{test}

\item{cache_dir}{test}
}
\description{
summarise_site
}
