% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_module.R
\name{tabulate_module}
\alias{tabulate_module}
\title{tabulate_module}
\usage{
tabulate_module(
  con = sidora.core::get_pandora_connection(),
  entity_type,
  entity_id = c(),
  filter_entity_type = NA,
  filter_string = NA,
  as_tsv = T,
  as_id_list = F,
  cache_dir = "~/.sidora"
)
}
\arguments{
\item{con}{test}

\item{entity_type}{test}

\item{filter_string}{test}

\item{as_tsv}{test}

\item{cache_dir}{test}
}
\description{
tabulate_module
}
